finalized rich text input
