+++ public/manifest.json
@@ -1,7 +1,7 @@
 {
   "manifest_version": 3,
   "name": "I Notes",
-  "version": "1.0",
+  "version": "1.1",
   "description": "A simple extension to take notes / developer commands / passwords / etc",
   "permissions": ["storage"],
   "action": {
